+++ config/functionNotFoundRule.neon
@@ -1,5 +1,5 @@
 services:
     -
-        class: PhpStanTwigAnalysis\Twig\Rule\FunctionNotFoundRule
+        class: PhpStanTwigAnalysis\Twig\Rule\FunctionNotFoundTwigRule
         tags:
             - twig_rule
