+++ tests/Twig/Rule/FilterNotFoundTwigRule/phpstan.neon
