+++ tests/Twig/Rule/ForbiddenFunctionsTwigRule/Fixtures/skip-template-uses-no-forbidden-functions.html.twig
