+++ tests/Twig/Rule/ForbiddenFunctionsTwigRule/Fixtures/template-uses-forbidden-functions.html.twig
