+++ tests/Twig/Rule/ForbiddenFunctionsTwigRule/phpstan.neon
