+++ tests/Twig/Rule/ForbiddenVariablesTwigRule/Fixtures/skip-template-uses-no-forbidden-variables.html.twig
