+++ tests/Twig/Rule/ForbiddenVariablesTwigRule/Fixtures/template-uses-forbidden-variables.html.twig
