+++ tests/Twig/Rule/ForbiddenVariablesTwigRule/phpstan.neon
