+++ tests/Twig/Rule/FunctionNotFoundTwigRule/phpstan.neon
