+++ tests/Twig/Rule/LawOfDemeterTwigRule/Fixtures/template-uses-more-than-the-allowed-number-of-dots.html.twig
