+++ tests/Twig/Rule/LawOfDemeterTwigRule/phpstan.neon
