+++ tests/Twig/Rule/NoDynamicIncludesTwigRule/Fixtures/skip-template-uses-constant-include.html.twig
