+++ tests/Twig/Rule/NoDynamicIncludesTwigRule/Fixtures/template-uses-dynamic-include.html.twig
