+++ tests/Twig/Rule/NoDynamicIncludesTwigRule/phpstan.neon
